+++ src/main/java/roomescape/admin/AdminController.java
@@ -4,7 +4,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import roomescape.reservation.ReservationService;
+import roomescape.reservation.service.ReservationService;
 
 @Controller
 @RequestMapping("/admin")
