+++ src/main/java/roomescape/reservation/dto/ReservationResponse.java
@@ -2,7 +2,7 @@ package roomescape.reservation.dto;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
-import roomescape.reservation.Reservation;
+import roomescape.reservation.domain.Reservation;
 
 public record ReservationResponse(long id, String name, LocalDate date, LocalTime time) {
     public static ReservationResponse from(Reservation reservation) {
