+++ src/main/java/roomescape/reservation/repository/ReservationRepository.java
@@ -2,7 +2,7 @@ package roomescape.reservation.repository;
 
 import java.util.List;
 import org.springframework.stereotype.Repository;
-import roomescape.reservation.Reservation;
+import roomescape.reservation.domain.Reservation;
 
 @Repository
 public interface ReservationRepository {
