+++ src/main/java/roomescape/reservation/service/ReservationService.java
@@ -1,7 +1,8 @@
-package roomescape.reservation;
+package roomescape.reservation.service;
 
 import java.util.List;
 import org.springframework.stereotype.Service;
+import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.dto.ReservationRequest;
 import roomescape.reservation.dto.ReservationResponse;
 import roomescape.reservation.repository.ReservationRepository;
